requirements.txt
beautifulsoup4==4.12.2
selenium==4.11.2
streamlit==1.26.0
chromedriver-binary-auto
openai